Abstracted deployment code into a job.
