+++ resources/views/project/deployment.blade.php
@@ -1,4 +1,4 @@
 <x-layout>
     <h1>Deployment</h1>
-    <livewire:poll-activity :activity="$activity" />
+    <livewire:poll-activity :activity="null" :deployment_uuid="$deployment_uuid" />
 </x-layout>
